move examples in src
